Move storybook helpers inside `src`
