+++ modules/react-components/.storybook/tsconfig.json
@@ -9,6 +9,6 @@
     "../../../node_modules/@nrwl/react/typings/cssmodule.d.ts",
     "../../../node_modules/@nrwl/react/typings/image.d.ts"
   ],
-  "exclude": ["../**/*.spec.ts" , "../**/*.spec.js", "../**/*.spec.tsx", "../**/*.spec.jsx"],
-  "include": ["../src/**/*", "*.js", "../src/*.stories.*" ]
+  "exclude": [ "../**/*.spec.ts" , "../**/*.spec.js", "../**/*.spec.tsx", "../**/*.spec.jsx" ],
+  "include": [ "../src/**/*", "*.js", "../src/*.stories.*" ]
 }
