+++ modules/react-components/src/components/brand/brand.stories.meta.ts
@@ -16,8 +16,8 @@
  * under the License.
  */
 
-import { StoryCategories } from "../../../storybook-helpers/hierarchy";
-import { StoryMetaInterface } from "../../../storybook-helpers/models";
+import { StoryCategories } from "../../storybook-helpers/hierarchy";
+import { StoryMetaInterface } from "../../storybook-helpers/models";
 
 export const meta: StoryMetaInterface = {
     components: [ "ProductBrand", "Logo" ],
